requirements.txt
# NEMWAS Framework Requirements
# Core NPU/OpenVINO support
openvino>=2024.0.0
openvino-genai>=2024.3
nncf>=2.7.0  # Neural Network Compression Framework

# LLM and ML frameworks
transformers>=4.36.0
torch>=2.1.0
tokenizers>=0.15.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Agent framework
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0

# Performance monitoring
prometheus-client>=0.19.0
psutil>=5.9.0
py-cpuinfo>=9.0.0

# Data handling
numpy>=1.24.0
pandas>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Plugin system
pluggy>=1.3.0
importlib-metadata>=6.8.0

# Logging and monitoring
structlog>=23.2.0
python-json-logger>=2.0.7

# Optional accelerators
onnx>=1.15.0
onnxruntime>=1.16.0
